# api/openapi.yaml
openapi: 3.0.0
info:
  title: Podman Desktop AI Lab API
  description: API for interacting with the Podman Desktop AI Lab service.
  version: 0.0.1
servers:
  - url: http://{host}:{port}
    description: Podman Desktop AI Lab API server
    variables:
      host:
        default: 127.0.0.1
      port:
        default: '10434'
...
paths:
...
  /api/pull:
    post:
      operationId: pullModel
      tags:
        - models
      description: |
        Download a model from the Podman AI Lab catalog.
      summary: |
        Download a model from the Podman AI Lab Catalog.
      requestBody:
        required: true
        description: Request to pull a model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullRequest'
      responses:
        '200':
          description: Model was successfully pulled
          content:
            application/x-ndjson:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
  /api/show:
    post:
      operationId: showModel
      tags:
        - models
      description: |
        Not implemented, returns an empty object - Show information about a model including details, modelfile, template, 
        parameters, license, and system prompt.
      summary: |
        Show information about a model including details, modelfile, template, 
        parameters, license, and system prompt.
      requestBody:
        required: true
        description: Request to show a model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowRequest'
      responses:
        '200':
          description: The model's information was successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
  /api/generate:
    post:
      operationId: generateResponse
      tags:
        - generate
      description: |
        Generate a response for a given prompt with a provided model. This is 
        a streaming endpoint, so there will be a series of responses. The 
        final response object will include statistics and additional data from 
        the request.
      summary: |
        Generate a response for a given prompt with a provided model. This is 
        a streaming endpoint, so there will be a series of responses. The final 
        response object will include statistics and additional data from the 
        request.
      requestBody:
        required: true
        description: Request to generate a response
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: A response was successfully generated for the prompt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
  /api/chat:
    post:
      operationId: generateChat
      tags:
        - chat
        - generate
      description: | 
        Generate the next message in a chat with a provided model. This is a 
        streaming endpoint, so there will be a series of responses. Streaming 
        can be disabled using "stream": false. The final response object will 
        include statistics and additional data from the request.
      summary: |
        Generate the next message in a chat with a provided model. This is a 
        streaming endpoint, so there will be a series of responses. Streaming 
        can be disabled using "stream": false. The final response object will 
        include statistics and additional data from the request.
      requestBody:
        required: true
        description: Request to generate a response in a chat
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: The next message was successfully generated for the chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
components:
  schemas:
...
    ModelDetails:
      type: object
      description: Details about a model
      properties:
        parent_model:
          type: string
        format:
          type: string
        family:
          type: string
        families:
          type: array
          items:
            type: string
        parameter_size:
          type: string
        quantization_level:
          type: string
    PullRequest:
      type: object
      description: Request to pull a model
      properties:
        model:
          type: string
          description: The name of the model to pull
          example: instructlab/granite-7b-lab-GGUF
        insecure:
          type: boolean
          description: |
            allow insecure connections to the catalog.
        stream:
          type: boolean
          description: |
            If false the response will be returned as a single response object,
            rather than a stream of objects
      required:
        - model
    ProgressResponse:
      type: object
      description: The response returned from various streaming endpoints
      properties:
        status:
          type: string
          description: The status of the request
        digest:
          type: string
          description: The SHA256 digest of the blob
        total:
          type: integer
          description: The total size of the task
        completed:
          type: integer
          description: The completed size of the task
    ShowRequest:
      type: object
      description: Request to show a model
      properties:
        model:
          type: string
          description: The name of the model to show
      required:
        - model
    ShowResponse:
      type: object
      description: Response from a show request
      properties:
        license:
          type: string
          description: The model license
        modelfile:
          type: string
          description: The modelfile content
        parameters:
          type: string
          description: The model parameters
        template:
          type: string
          description: The model template
        system:
          type: string
          description: The model system message/prompt
        details:
          $ref: '#/components/schemas/ModelDetails'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    GenerateRequest:
      type: object
      description: Request to generate a response
      properties:
        model:
          type: string
          description: The model name
        prompt:
          type: string
          description: The prompt to generate a response for
        suffix:
          type: string
        images:
          type: array
          items:
            type: string
            format: byte
          description: |
            A list of base64-encoded images (for multimodal models such as 
            llava)        
        format:
          type: string
          description: |
            The format to return a response in. Currently the only accepted 
            value is json
        system:
          type: string
          description: |
            System message to (overrides what is defined in the Modelfile)
        template:
          type: string
          description: |
            The prompt template to use (overrides what is defined in the 
            Modelfile)
        context:
          type: array
          items:
            type: integer
          description: |
            The context parameter returned from a previous request to generate, 
            this can be used to keep a short conversational memory
          example: []
        stream:
          type: boolean
          description: |
            If false the response will be returned as a single response object, 
            rather than a stream of objects
        raw:
          type: boolean
          description: |
            If true no formatting will be applied to the prompt. You may choose 
            to use the raw parameter if you are specifying a full templated 
            prompt in your request to the API
        keep_alive:
          $ref: '#/components/schemas/Duration'
      required:
        - model
    GenerateResponse:
      type: object
      description: Response from a generate request
      properties:
        model:
          type: string
          description: The model name that generated the response
        created_at:
          type: string
          format: date-time
          description: Timestamp of the response
        response:
          type: string
          description: |
            The textual response itself. When done, empty if the response was 
            streamed, if not streamed, this will contain the full response
        done:
          type: boolean
          description: Specifies if the response is complete
        context:
          type: array
          items:
            type: integer
          description: |
            When done, encoding of the conversation used in this response
        total_duration:
          type: number
          description: When done, time spent generating the response
        load_duration:
          type: number
          description: When done, time spent in nanoseconds loading the model
        prompt_eval_count:
          type: integer
          description: When done, number of tokens in the prompt
        prompt_eval_duration:
          type: number
          description: |
            When done, time spent in nanoseconds evaluating the prompt
        eval_count:
          type: integer
          description: When done, number of tokens in the response
        eval_duration:
          type: number
          description: |
            When done, time in nanoseconds spent generating the response 
    ChatRequest:
      type: object
      description: Request to generate a response in a chat
      properties:
        model:
          type: string
          description: The model name
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Messages of the chat - can be used to keep a chat memory
        stream:
          type: boolean
          description: Enable streaming of returned response
        format:
          type: string
          description: Format to return the response in (e.g. "json")
        keep_alive:
          $ref: '#/components/schemas/Duration'
        options:
          $ref: '#/components/schemas/Options'
    ChatResponse:
      type: object
      description: Response from a chat request
      properties:
        model:
          type: string
          description: The model name
        created_at:
          type: string
          format: date-time
          description: Timestamp of the response
        message:
          $ref: '#/components/schemas/Message'
        done_reason:
          type: string
          description: Reason the model stopped generating text
        done:
          type: boolean
          description: Specifies if the response is complete
        total_duration:
          type: number
          description: Total duration of the request
        load_duration:
          type: number
          description: Load duration of the request
        prompt_eval_count:
          type: integer
          description: Count of prompt evaluations
        prompt_eval_duration:
          type: number
          description: Duration of prompt evaluations
        eval_count:
          type: integer
          description: Count of evaluations
        eval_duration:
          type: number
          description: Duration of evaluations
    Message:
      type: object
      description: A message in a chat
      properties:
        role:
          type: string
        content:
          type: string
        images:
          type: array
          items:
            type: string
            format: byte
    Duration:
      type: string
      description: A string representing the duration
      example: "5m"
...